Add url cache busting
Since the charge tool is intended to be modifying objects in S3, we
don't want our browser caching those objects. Prepending the url with a
"random" query string (the curret timestamp) means we'll never have a
browser cached version of that object.

Use this new `uncached_url` where it's appropriate.