Move existence check to after prepare
If we're converting an uploaded file to JPEG, we change the upload
destination by munging the file extension. However, we were checking the
destination _before_ changing the extension. This is a bug.

Move the S3 existence check to _after_ `prepare_source_file`, which
munges the file extension. This ensures that we're checking the correct
intended upload destination.