Add an edit link on the upload page
Very frequently, I find myself uploading an image and immediately editing it. This provides a convenient edit link directly on the new image, rather than having to click to the image view to see the link.

In the process, I also switched the string quotes on the view link to avoid the need to escape the double-quotes.

This is currently untested.