cache object references
This was the slow-down. Each call to these properties on `Asset` was
making the S3 `head-object` API call over and over again. We can cache
the result. This reduces the S3 calls down from 18 per view to 2!

This is safe because `Asset` object references are specific to each run
of a route handler. Ref the main `charge.rb` route list.